server/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
accelerate
librosa
soundfile
numpy
scipy
python-jose[cryptography]
bcrypt<4.0.0
python-multipart
passlib
